[libromdata] Nintendo3DSPrivate: Reduce `headers_loaded` from uint32_t to uint8_t.
Only 5 bits are currently used, and this reduces padding due to 64-bit
alignment in the mxh union.

Total struct size before: 9,176
Total struct size after:  9,168
Difference:                  -8

+++ src/libromdata/Handheld/Nintendo3DS_p.hpp
@@ -80,7 +80,7 @@ public:
 		HEADER_TMD	= (1U << 3),	// ticket, tmd_header
 		HEADER_NCSD	= (1U << 4),	// ncsd_header, cinfo_header
 	};
-	uint32_t headers_loaded;	// HeadersPresent
+	uint8_t headers_loaded;	// HeadersPresent
 
 	// Media unit shift
 	// This is usually 9 (512 bytes), though NCSD images
